Remove duplication with ItemBox
